requirements.txt
# Core
httpx==0.27.0
beautifulsoup4==4.12.3
anthropic==0.25.0
slack-sdk==3.27.1
apscheduler==3.10.4
python-dotenv==1.0.1

# Dashboard API
flask==3.0.2
flask-cors==4.0.0

# Production WSGI server
gunicorn==22.0.0

# JavaScript rendering (optional — only needed for js_render=true competitors)
# Install with: pip install playwright && playwright install chromium
playwright==1.42.0

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-mock==3.12.0
